feat: remove components
Remove both the sponsors and services pages.
