+++ src/app/app-routing.module.ts
@@ -4,19 +4,14 @@ import { HomeComponent } from './home/home.component';
 import { ProjectsComponent } from './projects/projects.component';
 import { CertificationsComponent } from './certifications/certifications.component';
 import { ConnectComponent } from './connect/connect.component';
-import { CalendarComponent } from './calendar/calendar.component';
 import { ContactComponent } from './contact/contact.component';
-import { SponsorsComponent } from './sponsors/sponsors.component';
 
 const routes: Routes = [
-  { path: 'home', component: HomeComponent },
+  { path: '', pathMatch: 'full', component: HomeComponent },
   { path: 'projects', component: ProjectsComponent },
   { path: 'certifications', component: CertificationsComponent },
   { path: 'connect', component: ConnectComponent },
-  { path: 'services', component: CalendarComponent },
   { path: 'contact', component: ContactComponent },
-  { path: 'sponsors', component: SponsorsComponent },
-  { path: '', redirectTo: '/home', pathMatch: 'full' },
 ];
 
 @NgModule({
