--- src/app/calendar/calendar.component.css
@@ -1,18 +1,0 @@
-* {
-  text-align: center;
-}
-
-ul {
-  list-style-position: inside;
-  list-style-type: "🙂";
-}
-
-.text-content {
-  max-width: 750px;
-  width: 90vw;
-  margin: auto;
-  background-color: rgba(0, 0, 0, 0.75);
-  backdrop-filter: blur(10px);
-  border-radius: 10px;
-  margin-top: 10px;
-}
