--- src/app/calendar/calendar.component.html
@@ -1,50 +1,0 @@
-<div class="text-content">
-  <h1>Calendar</h1>
-  <p>
-    I offer a few different services related to development. Meetings can be
-    conducted through Discord, Google Meet, Skype, or another platform if you
-    prefer.
-  </p>
-  <h2>Freelance App Development</h2>
-  <p>
-    Do you need a full webpage or app built for you? I am more than happy to
-    provide that service. Please note that pricing varies depending on the size
-    and scope of the application you need built, as well as extra services such
-    as hosting.
-  </p>
-  <h2>Pair Programming Sessions</h2>
-  <p>
-    We can spend some time working on one of your projects together. With this
-    service, I will work with you directly in a 1-on-1 environment to write code
-    for your project together. Unlike other services, I will be directly writing
-    code with you and contributing to your project.
-  </p>
-  <h2>Code Review</h2>
-  <p>
-    If you are stuck on a certain bug, or aren't sure if your code is the most
-    readable or maintainable, this is the service for you! This is a 1-on-1
-    session where I will walk through your code with you, identify potential
-    concerns, and help resolve issues. While I will not be writing code for you,
-    I will gladly talk you through some potential approaches.
-  </p>
-  <h2>Mentoring Session</h2>
-  <p>
-    If you need some guidance on your development journey, whether it's
-    determining the next steps of your learning path, seeking advice on your job
-    search, or general coding questions, this is your chance. This service is a
-    1-on-1 meeting to talk through any questions or concerns you are having.
-  </p>
-  <p>
-    Our initial meeting will be a consultation, to determine if I am able to
-    meet your needs. If you would like to schedule a consultation, use the
-    calendar below:
-  </p>
-</div>
-
-<!-- Calendly inline widget begin -->
-<div
-  class="calendly-inline-widget"
-  data-url="https://calendly.com/nhcarrigan/meeting-request?background_color=3a3240&text_color=dfc0fa&primary_color=57426f"
-  style="min-width: 320px; height: 630px"
-></div>
-<!-- Calendly inline widget end -->
