+++ src/app/contact/contact.component.html
@@ -1,7 +1,10 @@
 <div class="text-content">
   <h1>Contact me!</h1>
-  <p>Please fill out the below form with as much detail as possible. I will
-    return your message as soon as I am able!</p>
+  <p>
+    Please fill out the below form with as much detail as possible. I will
+    return your message as soon as I am able! If you prefer to schedule a
+    meeting, you may use the calendar widget.
+  </p>
   <form action="https://formspree.io/f/xvowlnwe" method="POST">
     <label>
       Your name:
@@ -17,8 +20,21 @@
     </label>
     <label>
       Message:
-      <textarea class="custom-input" name="message" cols=30 rows=20></textarea>
+      <textarea
+        class="custom-input"
+        name="message"
+        cols="30"
+        rows="20"
+      ></textarea>
     </label>
     <button class="custom-btn" type="submit">Send</button>
   </form>
 </div>
+
+<!-- Calendly inline widget begin -->
+<div
+  class="calendly-inline-widget"
+  data-url="https://calendly.com/nhcarrigan/meeting-request?background_color=3a3240&text_color=dfc0fa&primary_color=57426f"
+  style="min-width: 320px; height: 630px"
+></div>
+<!-- Calendly inline widget end -->
