+++ src/app/header/header.component.html
@@ -6,11 +6,14 @@
   <p>Drinkin' coffee, boppin' tunes, bangin' code.</p>
   <app-audio></app-audio>
   <hr />
-  <p><a routerLink="/home" routerLinkActive="active-link">Home</a></p>
+  <p><a routerLink="/" routerLinkActive="active-link" [routerLinkActiveOptions]="{ exact: true }">Home</a></p>
   <p><a routerLink="/projects" routerLinkActive="active-link">Projects</a></p>
-  <p><a routerLink="/certifications" routerLinkActive="active-link">Certifications</a></p>
+  <p>
+    <a routerLink="/certifications" routerLinkActive="active-link"
+      >Certifications</a
+    >
+  </p>
   <p><a routerLink="/connect" routerLinkActive="active-link">Connect</a></p>
-  <p><a routerLink="/services" routerLinkActive="active-link">Services</a></p>
   <p><a routerLink="/contact" routerLinkActive="active-link">Contact</a></p>
   <p>
     <a
@@ -20,7 +23,6 @@
       >Chat</a
     >
   </p>
-  <p><a routerLink="/sponsors">Sponsors</a></p>
   <hr />
   <p>&copy; 2021</p>
   <p>
