--- src/app/sponsors/sponsors.component.css
@@ -1,46 +1,0 @@
-* {
-  text-align: center;
-}
-
-.text-content {
-  max-width: 750px;
-  width: 90vw;
-  margin: auto;
-  background-color: rgba(0, 0, 0, 0.6);
-  backdrop-filter: blur(10px);
-  border-radius: 10px;
-  margin-top: 10px;
-}
-
-a {
-  color: #dfc0fa;
-  transition: all .5s ease-in-out;
-}
-
-a:hover {
-  color: #3a3240;
-  background-color: #dfc0fa;
-}
-
-.custom-btn {
-  padding: 5px;
-  border: 2px solid #3a3240;
-  border-radius: 10px;
-  color: #3a3240;
-  background-color: #dfc0fa;
-  transition: all .5s ease-in-out;
-}
-
-.custom-btn:hover {
-  color: #dfc0fa;
-  background-color: #3a3240;
-  border: 2px solid #dfc0fa;
-}
-
-ul {
-  margin: auto;
-}
-
-li {
-  list-style-type: none;
-}
