--- src/app/sponsors/sponsors.component.html
@@ -1,66 +1,0 @@
-<div class="text-content">
-  <h1>My Sponsors!</h1>
-  <p>These are the amazing people who have sponsored my work.</p>
-  <p>
-    Would you like to become a monthly sponsor?
-    <a
-      class="custom-btn"
-      href="https://github.com/sponsors/nhcarrigan"
-      target="_blank"
-      rel="noopener noreferrer"
-      ><i class="fab fa-github"></i> Sponsor on GitHub</a
-    >
-  </p>
-  <p>
-    I also gladly accept one-time donations!
-    <a
-      class="custom-btn"
-      href="https://ko-fi.com/nhcarrigan"
-      target="_blank"
-      rel="noopener noreferrer"
-      ><i class="fas fa-coffee"></i> Buy me a Coffee</a
-    >
-  </p>
-  <p>
-    Or you can send a donation through PayPal!
-    <a
-      class="custom-btn"
-      href="https://www.paypal.com/cgi-bin/webscr?cmd=_donations&business=MA23ALWCRJ3MU&item_name=Support+my+development+projects&currency_code=USD"
-      target="_blank"
-      rel="noopener noreferrer"
-      ><i class="fab fa-paypal"></i> Make a Donation
-    </a>
-  </p>
-  <h2>Platinum Sponsors! ($100/month)</h2>
-  <ul>
-    <li class="platinum" *ngFor="let sponsor of platinum">
-      <a href={{sponsor.link}} target="_blank" rel="noopener noreferrer">
-        {{sponsor.name}} - {{sponsor.date}}
-      </a>
-    </li>
-  </ul>
-  <h2>Gold Sponsors! ($50/month)</h2>
-  <ul>
-    <li class="gold" *ngFor="let sponsor of gold">
-      <a href={{sponsor.link}} target="_blank" rel="noopener noreferrer">
-        {{sponsor.name}} - {{sponsor.date}}
-      </a>
-    </li>
-  </ul>
-  <h2>Silver Sponsors! ($25/month)</h2>
-  <ul>
-    <li class="silver" *ngFor="let sponsor of silver">
-      <a href={{sponsor.link}} target="_blank" rel="noopener noreferrer">
-        {{sponsor.name}} - {{sponsor.date}}
-      </a>
-    </li>
-  </ul>
-  <h2>Bronze Sponsors! ($10/month)</h2>
-  <ul>
-    <li class="bronze" *ngFor="let sponsor of bronze">
-      <a href={{sponsor.link}} target="_blank" rel="noopener noreferrer">
-        {{sponsor.name}} - {{sponsor.date}}
-      </a>
-    </li>
-  </ul>
-</div>
